refactor: 按功能拆分用户相关接口
